implemented part of review suggestions
