+++ common/src/main/scala/org/neo4j/spark/util/DriverCache.scala
@@ -6,6 +6,11 @@ import org.neo4j.spark.util.DriverCache.{cache, jobIdCache}
 import java.util.concurrent.ConcurrentHashMap
 import java.util.{Collections, function}
 
+object DriverCache {
+  private val cache: ConcurrentHashMap[Neo4jDriverOptions, Driver] = new ConcurrentHashMap[Neo4jDriverOptions, Driver]
+  private val jobIdCache = Collections.newSetFromMap[String](new ConcurrentHashMap[String, java.lang.Boolean]())
+}
+
 class DriverCache(private val options: Neo4jDriverOptions, private val jobId: String) extends Serializable with AutoCloseable {
   def getOrCreate(): Driver = {
     this.synchronized {
@@ -27,9 +32,4 @@ class DriverCache(private val options: Neo4jDriverOptions, private val jobId: St
       }
     }
   }
-}
-
-object DriverCache {
-  private val cache: ConcurrentHashMap[Neo4jDriverOptions, Driver] = new ConcurrentHashMap[Neo4jDriverOptions, Driver]
-  private val jobIdCache = Collections.newSetFromMap[String](new ConcurrentHashMap[String, java.lang.Boolean]())
 }
